refactor(models): move AttackGraph out of the asset package into models/attack.py (#161)
AttackGraph is not an OAM asset - it is the Sheyner-style bundle that wraps
the OAM shapes (the OA's recon output). academic-grounding.md already names
the AttackGraph / AttackTree / AttackForest trio as living in models/attack.py;
it was the odd one out, defined in the asset layer. Move it home.

- AttackGraph -> models/attack.py, placed first (OA describes the graph -> VR
  finds the trees -> PT searches the forest), composing the asset shapes via
  'from models.asset import ...'. attack now depends on asset (acyclic: no
  asset submodule imports attack; finding / h1 don't either).
- Drop models/asset/graph.py; models.asset/__init__ and its docstring no
  longer claim AttackGraph. models/__init__ re-exports it from models.attack.
- TestAttackGraph relocated into tests/models/test_attack.py alongside the
  AttackTree / AttackForest tests.

from models import AttackGraph is unchanged; the only direct importer was the
package __init__ itself. Pure relocation - no shape or behaviour change.
